+++ hcloud/certificate.go
@@ -85,6 +85,9 @@ func (l CertificateListOpts) values() url.Values {
 }
 
 // List returns a list of Certificates for a specific page.
+//
+// Please note that filters specified in opts are not taken into account
+// when their value corresponds to their zero value or when they are empty.
 func (c *CertificateClient) List(ctx context.Context, opts CertificateListOpts) ([]*Certificate, *Response, error) {
 	path := "/certificates?" + opts.values().Encode()
 	req, err := c.client.NewRequest(ctx, "GET", path, nil)