+++ hcloud/datacenter.go
@@ -83,6 +83,9 @@ func (l DatacenterListOpts) values() url.Values {
 }
 
 // List returns a list of datacenters for a specific page.
+//
+// Please note that filters specified in opts are not taken into account
+// when their value corresponds to their zero value or when they are empty.
 func (c *DatacenterClient) List(ctx context.Context, opts DatacenterListOpts) ([]*Datacenter, *Response, error) {
 	path := "/datacenters?" + opts.values().Encode()
 	req, err := c.client.NewRequest(ctx, "GET", path, nil)