+++ hcloud/network.go
@@ -122,6 +122,9 @@ func (l NetworkListOpts) values() url.Values {
 }
 
 // List returns a list of networks for a specific page.
+//
+// Please note that filters specified in opts are not taken into account
+// when their value corresponds to their zero value or when they are empty.
 func (c *NetworkClient) List(ctx context.Context, opts NetworkListOpts) ([]*Network, *Response, error) {
 	path := "/networks?" + opts.values().Encode()
 	req, err := c.client.NewRequest(ctx, "GET", path, nil)