+++ hcloud/server_type.go
@@ -103,6 +103,9 @@ func (l ServerTypeListOpts) values() url.Values {
 }
 
 // List returns a list of server types for a specific page.
+//
+// Please note that filters specified in opts are not taken into account
+// when their value corresponds to their zero value or when they are empty.
 func (c *ServerTypeClient) List(ctx context.Context, opts ServerTypeListOpts) ([]*ServerType, *Response, error) {
 	path := "/server_types?" + opts.values().Encode()
 	req, err := c.client.NewRequest(ctx, "GET", path, nil)