+++ hcloud/volume.go
@@ -106,6 +106,9 @@ func (l VolumeListOpts) values() url.Values {
 }
 
 // List returns a list of volumes for a specific page.
+//
+// Please note that filters specified in opts are not taken into account
+// when their value corresponds to their zero value or when they are empty.
 func (c *VolumeClient) List(ctx context.Context, opts VolumeListOpts) ([]*Volume, *Response, error) {
 	path := "/volumes?" + opts.values().Encode()
 	req, err := c.client.NewRequest(ctx, "GET", path, nil)